# 8265029: Preserve SIZED characteristics on slice operations (skip, limit)