--- a/src/java.base/share/classes/java/util/stream/AbstractPipeline.java
+++ b/src/java.base/share/classes/java/util/stream/AbstractPipeline.java
@@ -469,10 +469,10 @@ abstract class AbstractPipeline<E_IN, E_OUT, S extends BaseStream<E_OUT, S>>
         int flags = getStreamAndOpFlags();
         long size = StreamOpFlag.SIZED.isKnown(flags) ? spliterator.getExactSizeIfKnown() : -1;
         // Currently, we have no stateless SIZE_ADJUSTING intermediate operations,
-        // so we can simply ignore SIZE_ADJUSTING in parallel more, as adjustments
+        // so we can simply ignore SIZE_ADJUSTING in parallel streams, since adjustments
         // are already accounted in the input spliterator.
         //
-        // If we ever have SIZE_ADJUSTING intermediate operation,
+        // If we ever have a stateless SIZE_ADJUSTING intermediate operation,
         // we would need step back until depth == 0, then call exactOutputSize() for
         // the subsequent stages.
         if (size != -1 && StreamOpFlag.SIZE_ADJUSTING.isKnown(flags) && !isParallel()) {