--- a/src/java.base/share/classes/java/util/stream/PipelineHelper.java
+++ b/src/java.base/share/classes/java/util/stream/PipelineHelper.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012, 2015, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2012, 2021, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -82,7 +82,8 @@ abstract class PipelineHelper<P_OUT> {
      * The exact output size is known if the {@code Spliterator} has the
      * {@code SIZED} characteristic, and the operation flags
      * {@link StreamOpFlag#SIZED} is known on the combined stream and operation
-     * flags.
+     * flags. The exact output size may differ from spliterator size,
+     * if pipeline contains a slice operation.
      *
      * @param spliterator the spliterator describing the relevant portion of the
      *        source data