--- a/src/java.base/share/classes/java/util/stream/StreamOpFlag.java
+++ b/src/java.base/share/classes/java/util/stream/StreamOpFlag.java
@@ -325,12 +325,21 @@ enum StreamOpFlag {
      */
     // 12, 0x01000000
     SHORT_CIRCUIT(12,
-                  set(Type.OP).set(Type.TERMINAL_OP));
+                  set(Type.OP).set(Type.TERMINAL_OP)),
 
+    /**
+     * Characteristic value signifying that an operation may adjust the
+     * total size of the stream.
+     * <p>
+     * An intermediate operation can preserve or inject this value.
+     */
+    // 13, 0x04000000
+    SIZE_ADJUSTING(13,
+                   set(Type.OP));
+    
     // The following 2 flags are currently undefined and a free for any further
     // stream flags if/when required
     //
-    // 13, 0x04000000
     // 14, 0x10000000
     // 15, 0x40000000
 
@@ -628,6 +637,11 @@ enum StreamOpFlag {
      * The bit value to inject {@link #SHORT_CIRCUIT}.
      */
     static final int IS_SHORT_CIRCUIT = SHORT_CIRCUIT.set;
+
+    /**
+     * The bit value to inject {@link #SIZE_ADJUSTING}.
+     */
+    static final int IS_SIZE_ADJUSTING = SIZE_ADJUSTING.set;
 
     private static int getMask(int flags) {
         return (flags == 0)