--- a/src/java.base/share/classes/java/util/stream/StreamSpliterators.java
+++ b/src/java.base/share/classes/java/util/stream/StreamSpliterators.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012, 2017, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2012, 2021, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -217,19 +217,17 @@ class StreamSpliterators {
 
         @Override
         public final long estimateSize() {
-            init();
+            long exactSizeIfKnown = getExactSizeIfKnown();
             // Use the estimate of the wrapped spliterator
             // Note this may not be accurate if there are filter/flatMap
             // operations filtering or adding elements to the stream
-            return spliterator.estimateSize();
+            return exactSizeIfKnown == -1 ? spliterator.estimateSize() : exactSizeIfKnown;
         }
 
         @Override
         public final long getExactSizeIfKnown() {
             init();
-            return StreamOpFlag.SIZED.isKnown(ph.getStreamAndOpFlags())
-                   ? spliterator.getExactSizeIfKnown()
-                   : -1;
+            return ph.exactOutputSizeIfKnown(spliterator);
         }
 
         @Override