… test/jdk/java/util/stream/test/org/openjdk/tests/java/util/stream/CountTest.java
… test/jdk/java/util/stream/test/org/openjdk/tests/java/util/stream/SliceOpTest.java
… test/micro/org/openjdk/bench/java/util/stream/ops/ref/SliceToList.java
… test/micro/org/openjdk/bench/java/util/stream/ops/value/SizedCount.java
… test/micro/org/openjdk/bench/java/util/stream/ops/value/SizedSum.java
… test/micro/org/openjdk/bench/java/util/stream/ops/value/SliceToArray.java